changes heroku references to bitbucket
