+++ Gemfile
@@ -1,4 +1,4 @@
 source 'https://rubygems.org'
 
-# Specify your gem's dependencies in github_heroku_deployer.gemspec
+# Specify your gem's dependencies in github_bitbucket_deployer.gemspec
 gemspec
