+++ lib/github_bitbucket_deployer/exceptions.rb
@@ -1,4 +1,4 @@
-module GithubHerokuDeployer
+module GithubBitbucketDeployer
   class ConfigurationException < StandardError; end
   class CommandException < StandardError; end
 end
