+++ lib/github_bitbucket_deployer/version.rb
@@ -1,0 +1,3 @@
+module GithubBitbucketDeployer
+  VERSION = "0.4.1"
+end
