--- lib/github_heroku_deployer/version.rb
@@ -1,3 +1,0 @@
-module GithubHerokuDeployer
-  VERSION = "0.4.1"
-end
