+++ spec/lib/github_bitbucket_deployer/bitbucket_spec.rb
