+++ spec/lib/github_bitbucket_deployer/configuration_spec.rb
