+++ spec/lib/github_bitbucket_deployer/git_spec.rb
