+++ spec/lib/github_bitbucket_deployer_spec.rb
